Refactor/constants and types (#278)
* refactor: refactored types and constants in AudioArray

* refactor: refactored types and constants in AnalyserNode

* refactor: refactored types and constants in AudioBus

* fix: added necessary include

* refactor: refactored types and constants in AudioBuffer

* refactor: refactored types and constants in AudioBufferSourceNode

* refactor: refactored types and constants in BaseAudioContext

* refactor: refactored types and constants in AudioContext

* refactor: refactored types and constants in AudioDecoder and AudioDestinationNode

* refactor: refactored types and constants in AudioNode

* refactor: refactored types and constants in AudioNodeManager

* refactor: refactored types and constants in AudioParam

* refactor: refactored types and constants in AudioScheduledSourceNode

* refactor: refactored types and constants in BiquadFilterNode

* refactor: refactored types and constants in GainNode

* refactor: refactored types and constants in OscillatorNode

* refactor: refactored types and constants in StereoPannerNode

* refactor: refactored types and constants in PeriodicWave

* refactor: refactored types in utils

* ci: yarn format

* fix: fixed analyser types

* refactor: host objects

---------

+++ packages/react-native-audio-api/android/src/main/cpp/core/AudioPlayer.h
@@ -16,9 +16,9 @@ class AudioPlayer : public AudioStreamDataCallback {
   explicit AudioPlayer(const std::function<void(AudioBus *, int)> &renderAudio);
   AudioPlayer(
       const std::function<void(AudioBus *, int)> &renderAudio,
-      int sampleRate);
+      float sampleRate);
 
-  [[nodiscard]] int getSampleRate() const;
+  [[nodiscard]] float getSampleRate() const;
   void start();
   void stop();
 
@@ -32,7 +32,7 @@ class AudioPlayer : public AudioStreamDataCallback {
   std::shared_ptr<AudioStream> mStream_;
   std::shared_ptr<AudioBus> mBus_;
   bool isInitialized_ = false;
-  int sampleRate_;
+  float sampleRate_;
 };
 
 } // namespace audioapi
